requirements.txt
# DocuExtract - Odia Handwritten OCR
# Python Dependencies for Dataset Preparation and Model Training

# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Image processing
Pillow>=9.4.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data processing
numpy>=1.24.0,<2.0.0  # Compatible with scipy
pandas>=2.0.0

# Dataset tools
datasets>=2.14.0
huggingface-hub>=0.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
argparse

# Utilities
pathlib
json5>=0.9.0

# Optional: For advanced features
# albumentations>=1.3.0  # Advanced image augmentation
# tensorboard>=2.13.0    # Training visualization
# wandb>=0.15.0          # Experiment tracking

# Development tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# jupyter>=1.0.0
# ipykernel>=6.25.0
